Terprint.Python/AzureEventhouse/test/requirements.txt
azure-core-tracing-opentelemetry
azure-identity
azure-kusto-data
azure-kusto-ingest
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
azure-search-documents
azure-storage-blob